+++ changes.md
@@ -2,6 +2,16 @@
 
 This document outlines the changes made between versions of the GoatForms library.
 
+## V1.0.2
+
+## New Features
+
+New methods have been added to the `AppUtils` class to facilitate event logging.
+
+## Updated Features
+
+The `GetAppDataPath()` method in the `AppUtils` class now takes your application name or relative path to your application data directory as an argument.
+
 ##V1.0.1
 
 ### New Features